Check for malicious files

This function checks for the presence of files with extensions known to be associated with malicious activities. The function can be used to screen zip files or individual files for these potentially dangerous file types.

Usage

check_for_malicious_files(files)

Arguments

files

A character vector of file paths. These can be paths to zip files or individual files.

Value

A logical value indicating if any of the files in the input have a malicious file extension. Returns `TRUE` if any malicious file is found, otherwise `FALSE`.

Examples

if (FALSE) { # \dontrun{
  check_for_malicious_files("path'(s)'/to/your/files")
  check_for_malicious_files(utils::unzip("path/to/your/file.zip", list = TRUE)$Name)
} # }