Download all data alternative

This function allows users to download all data rather than one data set at a time.

Usage

download_all(
  file_path = NULL,
  s3_key_id = NULL,
  s3_secret_key = NULL,
  s3_region = NULL,
  s3_bucket = NULL,
  callback = NULL
)

Arguments

file_path

location and name of the zip file to create.

s3_key_id

A character string representing the AWS S3 access key ID.

s3_secret_key

A character string representing the AWS S3 secret access key.

s3_region

A character string representing the AWS S3 region.

s3_bucket

A character string representing the AWS S3 bucket name.

callback

Prints if the download was a success.

Value

Any return objects from the downloads.

Examples

if (FALSE) { # \dontrun{
    download_all_data <- download_all(file_path = "your/path/file.zip",
                                      s3_key_id = "your_s3_key_id",
                                      s3_secret_key = "your_s3_secret_key",
                                      s3_region = "your_s3_region",
                                      s3_bucket = "your_s3_bucket",
                                      callback = NULL)
} # }